refactor: Group Software Index API structs
Preparatory work for moving the Software Index code into the main OpoLua project. This should ensure we don’t have namespace collisions with structs like `Image`, etc.
